Struct cranelift_egraph::BumpVec

pub struct BumpVec<T> { /* private fields */ }

A vector of T stored within a BumpArena.

This is something like a normal Vec, except that all accesses and updates require a separate borrow of the BumpArena. This, in turn, makes the Vec itself very compact: only three u32s (12 bytes). The BumpSlice variant is only two u32s (8 bytes) and is sufficient to reconstruct a slice, but not grow the vector.

The BumpVec does not implement Clone or Copy; it represents unique ownership of a range of indices in the arena. If dropped, those indices will be unavailable until the arena is freed. This is “fine” (it is normally how arena allocation works). To explicitly free and make available for some allocations, a very rudimentary reuse mechanism exists via BumpVec::free(arena). (The allocation path opportunistically checks the first range on the freelist, and can carve off a piece of it if larger than needed, but it does not attempt to traverse the entire freelist; this is a compromise between bump-allocation speed and memory efficiency, which also influences speed through cached-memory reuse.)

The type T should not have a Drop implementation. This typically means that it does not own any boxed memory, sub-collections, or other resources. This is important for the efficiency of the data structure (otherwise, to call Drop impls, the arena needs to track which indices are live or dead; the BumpVec itself cannot do the drop because it does not retain a reference to the arena). Note that placing a T with a Drop impl in the arena is still safe, because leaking (that is, never calling Drop::drop()) is safe. It is merely less efficient, and so should be avoided if possible.

Implementations

impl<T> BumpVec<T>

pub fn as_slice<'a>(&'a self, arena: &'a BumpArena<T>) -> &'a [T]

Returns a slice view of this BumpVec, given a borrow of the arena.

pub fn as_mut_slice<'a>( &'a mut self, arena: &'a mut BumpArena<T> ) -> &'a mut [T]

Returns a mutable slice view of this BumpVec, given a mutable borrow of the arena.

pub fn len(&self) -> usize

Returns the length of this vector. Does not require access to the arena.

pub fn cap(&self) -> usize

Returns the capacity of this vector. Does not require access to the arena.

pub fn reserve(&mut self, extra_len: usize, arena: &mut BumpArena<T>)

Reserve extra_len capacity at the end of the vector, reallocating if necessary.

pub fn push(&mut self, t: T, arena: &mut BumpArena<T>)

Push an item, growing the capacity if needed.

pub fn clone(&self, arena: &mut BumpArena<T>) -> BumpVec<T>where T: Clone,

Clone, if T is cloneable.

pub fn truncate(&mut self, len: usize)

Truncate the length to a smaller-or-equal length.

pub fn free(self, arena: &mut BumpArena<T>)

Consume the BumpVec and return its indices to a free pool in the arena.

pub fn freeze(self, arena: &mut BumpArena<T>) -> BumpSlice<T>

Freeze the capacity of this BumpVec, turning it into a slice, for a smaller struct (8 bytes rather than 12). Once this exists, it is copyable, because the slice will never be freed.