Struct cranelift_egraph::CtxHashMap

pub struct CtxHashMap<K, V> { /* private fields */ }

A HashMap that takes external context for all operations.

Implementations

impl<K, V> CtxHashMap<K, V>

pub fn new() -> Self

Create an empty hashmap.

pub fn with_capacity(capacity: usize) -> Self

Create an empty hashmap with pre-allocated space for the given capacity.

impl<K, V> CtxHashMap<K, V>

pub fn insert<Ctx: CtxEq<K, K> + CtxHash<K>>(
&mut self,
k: K,
v: V,
ctx: &Ctx,
uf: &mut UnionFind
) -> Option<V>

Insert a new key-value pair, returning the old value associated with this key (if any).

pub fn get<'a, Q, Ctx: CtxEq<K, Q> + CtxHash<Q> + CtxHash<K>>(
&'a self,
k: &Q,
ctx: &Ctx,
uf: &mut UnionFind
) -> Option<&'a V>

Look up a key, returning a borrow of the value if present.

pub fn entry<'a, Ctx: CtxEq<K, K> + CtxHash<K>>(
&'a mut self,
k: K,
ctx: &'a Ctx,
uf: &mut UnionFind
) -> Entry<'a, K, V>

Return an Entry cursor on a given bucket for a key, allowing for fetching the current value or inserting a new one.