Trait cranelift_egraph::Analysis

pub trait Analysis {
    type L: Language;
    type Value: Clone + Default;

    // Required methods
    fn for_node(
        &self,
        ctx: &Self::L,
        n: &<Self::L as Language>::Node,
        values: &SecondaryMap<Id, Self::Value>,
    ) -> Self::Value;
    fn meet(
        &self,
        ctx: &Self::L,
        v1: &Self::Value,
        v2: &Self::Value,
    ) -> Self::Value;
}

A trait that allows the aegraph to compute a property of each node as it is created.

Required Associated Types

type L: Language

type Value: Clone + Default

Required Methods

fn for_node( &self, ctx: &Self::L, n: &<Self::L as Language>::Node, values: &SecondaryMap<Id, Self::Value>, ) -> Self::Value

fn meet(&self, ctx: &Self::L, v1: &Self::Value, v2: &Self::Value) -> Self::Value